refactor python files and class names
