+++ Dockerfile
@@ -14,4 +14,4 @@ COPY requirements.txt requirements.txt
 RUN pip install -r requirements.txt
 
 COPY . .
-CMD ["python", "-u", "pv.py", "-v"]
+CMD ["python", "-u", "main.py", "-v"]
