+++ examples/docker-compose.yml
@@ -1,0 +1,112 @@
+version: '3'
+
+services:
+  influxdb:
+    container_name: influxdb
+    image: influxdb:2.2
+    restart: unless-stopped
+    volumes:
+      - ./data/influxdb2:/var/lib/influxdb2
+      - ./data/influxdb2-config:/etc/influxdb2
+    ports:
+      - "8086:8086"
+
+  influxdb_cli:
+    container_name: influxdb_cli
+    links:
+      - influxdb
+    image: influxdb:2.2
+    entrypoint: influx setup --bucket main -t my_secret_influx_token -o acme --username=admin --password=my_secret_passwd --host=http://influxdb:8086 -f
+    restart: on-failure:10
+    depends_on:
+      - influxdb
+
+  mosquitto:
+    container_name: mosquitto
+    image: eclipse-mosquitto:latest
+    restart: always
+    ports:
+       - "1883:1883"
+    volumes:
+      - ./data/mosquitto/config:/mosquitto/config
+      - ./data/mosquitto/data:/mosquitto/data
+      - ./data/mosquitto/log:/mosquitto/log
+
+  pyfusionsolar:
+    container_name: pyfusionsolardatarelay
+    image: jsprnl/pyfusionsolardatarelay:latest
+    #build: ./PyFusionSolarDataRelay
+    restart: unless-stopped
+    depends_on:
+      - influxdb
+      - mosquitto
+    environment:
+      - debug_mode=True
+      - fusionsolar_kiosk_site_name=inverter01
+
+      - fusionsolar_kiosk_enabled=True
+      - fusionsolar_kiosk_api_kkid=GET_THIS_STRING_FROM_YOUR_KIOSK_URL
+      - fusionsolar_kiosk_fetch_cron_minute=0,30 # Please note that the kiosk only seems to update each 30mins
+      
+      - influxdb_enabled=True
+      - influxdb_is_v2=True
+      - influxdb_host=influxdb
+      - influxdb_v2_protocol=http
+      - influxdb_v2_org=acme
+      - influxdb_v2_bucket=fusionsolar
+      - influxdb_v2_token=GENERATE_TOKEN_IN_INFLUXDB_ADMIN_PANEL
+
+      - pvoutput_enabled=True
+      - pvoutput_system_id=123456
+      - pvoutput_api_key=GENERATE_THIS_AND_SYSTEMID_ON_PVOUTPUT.ORG
+
+      - mqtt_enabled=True
+      - mqtt_host=mosquitto
+      - mqtt_auth=False
+
+      - meetdata_nl_enabled=False
+      - meetdata_nl_username=my@username.com
+      - meetdata_nl_password=secretpassword
+
+      - meetdata_nl_meter_sysname=transformer01
+      - meetdata_nl_meter_ean=000000000000000000
+      - meetdata_nl_meter_id=0000000000
+
+      - meetdata_nl_meter2_enabled=False
+      - meetdata_nl_meter2_sysname=transformer02
+      - meetdata_nl_meter2_ean=000000000000000000
+      - meetdata_nl_meter2_id=0000000000
+
+
+
+    volumes:
+      - /etc/localtime:/etc/localtime:ro
+
+  grafana:
+    depends_on:
+      - influxdb
+    image: grafana/grafana-oss:8.5.5
+    container_name: grafana
+    restart: unless-stopped
+    ports:
+      - 3000:3000
+    user: "1000:1000"
+    volumes:
+      - ./data/grafana:/var/lib/grafana
+    environment:
+      - GF_AUTH_ANONYMOUS_ENABLED=true
+      - GF_AUTH_ANONYMOUS_ORG_NAME=Main Org.
+      - GF_AUTH_ANONYMOUS_ORG_ROLE=Viewer
+      - GF_AUTH_ANONYMOUS_HIDE_VERSION=True
+      - GF_SECURITY_ALLOW_EMBEDDING=True
+
+  homeassistant:
+    container_name: homeassistant
+    depends_on:
+      - mosquitto
+    image: homeassistant/home-assistant
+    volumes:
+      - ./data/hass:/config
+      - /etc/localtime:/etc/localtime:ro
+    restart: unless-stopped
+    network_mode: host
