+++ modules/relay_fusionsolar.py
@@ -1,24 +1,22 @@
 import time
 from apscheduler.schedulers.blocking import BlockingScheduler
-from pvconfmodels import BaseConf
-from pvinflux import PvInflux
-from pvoutputorg import PvOutputOrg
-from pvfusionsolar import PvFusionSolar
-from pvmqtt import PvMqtt
+from modules.conf_models import BaseConf
+from modules.write_influxdb import WriteInfluxDb
+from modules.write_pvoutput import WritePvOutput
+from modules.fetch_fusionsolar_kiosk import FetchFusionSolarKiosk
+from modules.write_mqtt import WriteMqtt
 
-
-
-class PvRelay:
+class RelayFusionSolar:
     def __init__(self, conf: BaseConf, logger):
         self.conf = conf
         self.logger = logger
         self.logger.debug("PvRelay class instantiated")
 
-        self.pvfusionsolar = PvFusionSolar(conf, logger)
-        self.pvoutput = PvOutputOrg(conf, logger)
-        self.pvmqtt = PvMqtt(conf, logger)
+        self.pvfusionsolar = FetchFusionSolarKiosk(conf, logger)
+        self.pvoutput = WritePvOutput(conf, logger)
+        self.pvmqtt = WriteMqtt(conf, logger)
 
-        self.pvinflux = PvInflux(self.conf, self.logger)
+        self.pvinflux = WriteInfluxDb(self.conf, self.logger)
         self.pvinflux_initialized = False
 
         self.logger.info("Starting PvRelay on separate thread")
